Changes from Buzz to Guzzle
Replaces the Buzz Browser with Guzzle. Writes basic service descriptions
for Board Game Geek's XML API 2. Adds clients for XML API and XML API 2.

Replaces existing functionality of getting a game and displaying the name
of the games with similar funcitionality in Guzzle.

Creates response classes that replace the factories and are in charge of
converting the XML data from BGG to the related objects in shelf.