Completes Basic Boardgame Model
The board game model now stores all of the basic information returned
from the API and has getters to return that information. Getting a
board game's expansions returns an array of board game models.

The only exception to the above is the poll information returned from
the API. This is not yet stored.