Adds Basic Board Game and Name Entities
Updates the board game factory to focus on only a single board game at
a time. Later, we will add a board game collection factory if you need
to work with an array.

Adds logic to board game factory to convert a board game xml item into
an array. Adds a fromArray factory method to convert a board game array
into a board game object. Updates fromBggXml to first convert item to
array and then use the fromArray method to return a board game entity.
Adds fromArray method to factory interface.

Adds basic methods to board game entity for getting the primary name of
a board game and checking if it is an expansion. Adds method to name
entity to report if it is the primary name and adds method to name
collection to find the primary name object.

Adds tests for board game and name entities and fixture files to be
used in populating the test objects.