Adds Methods for Sort by Names
Adds a method to the name entity to get its value starting at the sort
index. An optional flag passed to the argument controls whether the
characters before the start of the sort index are appended to the end
of the new sort value. Method defaults to appending.

Adds a sortByPrimaryName method to the BoardGameCollection and a
sortByValue to the NameCollection.

Adds tests for all new methods.