Adds Boardgame Link Collection and Entity
Creates classes for board game link collections and entities. All links
do not expand to their full entity types but rather serve as stubs. If
more information is needed, the id from the stub can be used to request
more information about a specific link.

Adds methods to the board game entity to get all the links of a subset
of a specific type of link. Adds similar methods to the link collection
entity to filter by the requested type.

Adds a filterByArray and a filterBy* method to the abstract data entity
collection to allow only returning a subset of an entire collection as
a new collection object.

Adds a get* method to the abstract data entity collection that returns
an array of that value for each child in the collection.

Moves some of the inflection helper classes in AbstractDataEntity out
to AbstractEntity and changed AbstractDataEntity and
AbstractDataEntityCollection to extend from AbstractEntity.

Adds related tests.