Adds hasLink and canFinishIn Methods
Adds methods to check if a game has a link with a specific value. Adds
methods to check for a link of specific value and type. Adds related
tests.

Adds methods to check if a game has expansions, implementations, or
compilations. Adds related tests.

Adds compilations as a valid link type and the necessary methods to
the linkCollection and boardgame entity classes along with the related
test methods.

Adds method to check if a game can finish within a specified time frame
and related tests for method.