Adds Poll Entities and Basic Getter Commands
Adds Poll and PollCollection Entities. Adds methods to the Boardgame
entity to either get the PollCollection or return a specific Poll by
name from the collection.

Adds similar methods to the PollCollection object to make it easier
to get named polls from the collection.

Adds related tests for the new methods.