Adds getName to Collection Entity
Adds a getName method to the Collection entity. This resulted in the
Boardgame Name entity having some methods move out to an AbstractName
entity which the Boardgame and Collection name entities then extend
from.

Adds tests for new classes and methods. Fixes an issue with getting
the sort name when the sort_index is 1.